Add files via upload
I performed end-to-end analysis on the Telecom Customer Churn dataset. It is hoped that this analysis will be insightful and actionable for the business. I performed several analyzes on the dataset using SQL, and carried out a thorough Diagnostic Analysis using Python to find problems that have not been identified, and solve them from the root of the problem.